Added the ability to open documents
Used Firebase + React Router to be able to open documents created in the database. Created a new type of page called DocEditor, which will open when you click on a document and will ultimately allow you to read its contents and edit them, though that has not been implemented yet. You can only open a new page corresponding to that document.