Laid framework for write permissions.
Set up documents so that when they're created, the current user will be noted as the "owner" of the document and only they will be able to write to it. Currently, login and sign up hasn't been set up. Therefore, when the current user is null documents will be created so that anybody can write to them.